+++ openstack/dcs/v1/configs/Update.go
@@ -1,0 +1,16 @@
+package configs
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func Update(client *golangsdk.ServiceClient, instanceID string, opts UpdateOptsBuilder) (r UpdateResult) {
+	b, err := opts.ToConfigUpdateMap()
+	if err != nil {
+		r.Err = err
+		return
+	}
+
+	_, r.Err = client.Put(rootURL(client, instanceID), b, nil, &golangsdk.RequestOpts{
+		OkCodes: []int{204},
+	})
+	return
+}
